ha_rocksdb.cc: update_side_plugin_dbopt: override by mysql conf
If a var is defined in any place, it will override which defined
in sideplugin json/yaml.

Just enum_variable_source::COMPILED indicate the var value is
from compile time, and not defined by any conf or cmd line.

sideplugin repo can not be loaded earlier when rocksdb_se is
not loaded, the problem is the register order of sideplugin(s)
defined in rocksdb_se is unspecified. So we init in 2 steps,
the first is in rdb_init_rocksdb_db_options(), the second is
update_side_plugin_dbopt().